requirements.txt
aiofiles==0.4.0
appdirs==1.4.3
async-exit-stack==1.0.1
async-generator==1.10
atomicwrites==1.3.0
attrs==19.2.0
black==19.3b0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
dataclasses==0.6
entrypoints==0.3
fastapi==0.42.0
flake8==3.7.8
gunicorn==19.9.0
h11==0.8.1
h2==3.1.1
hpack==3.0.0
hstspreload==2019.10.14
httptools==0.0.13
httpx==0.7.5
hyperframe==5.2.0
idna==2.8
importlib-metadata==0.23
Jinja2==2.10.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
motor==2.0.0
mypy==0.730
mypy-extensions==0.4.2
packaging==19.2
pluggy==0.13.0
py==1.8.0
pycodestyle==2.5.0
pydantic==0.32.2
pyflakes==2.1.1
pymongo==3.9.0
pyparsing==2.4.2
pytest==5.2.1
python-dotenv==0.10.3
rfc3986==1.3.2
sentry-sdk==0.13.0
six==1.12.0
starlette==0.12.9
toml==0.10.0
typed-ast==1.4.0
typing-extensions==3.7.4
uvicorn==0.9.0
uvloop==0.13.0
wcwidth==0.1.7
websockets==8.0.2
zipp==0.6.0
